Fix the dry run queries - we have to mock different objects now -
