+++ src/fides/api/api/v1/endpoints/privacy_request_endpoints.py
@@ -111,7 +111,7 @@ from fides.api.service.privacy_request.request_service import (
 from fides.api.task.create_tasks import log_task_queued
 from fides.api.task.execute_tasks import run_access_node, run_prerequisite_task_checks
 from fides.api.task.filter_results import filter_data_categories
-from fides.api.task.graph_task import EMPTY_REQUEST, collect_queries
+from fides.api.task.graph_task import EMPTY_REQUEST, EMPTY_REQUEST_TASK, collect_queries
 from fides.api.task.task_resources import TaskResources
 from fides.api.tasks import MESSAGING_QUEUE_NAME
 from fides.api.util.api_router import APIRouter
@@ -805,9 +805,12 @@ def get_request_preview_queries(
             k: "something" for k in dataset_graph.identity_keys.values()
         }
         traversal: Traversal = Traversal(dataset_graph, identity_seed)
+
         queries: Dict[CollectionAddress, str] = collect_queries(
             traversal,
-            TaskResources(EMPTY_REQUEST, Policy(), connection_configs, db),
+            TaskResources(
+                EMPTY_REQUEST, Policy(), connection_configs, EMPTY_REQUEST_TASK, db
+            ),
         )
         return [
             DryRunDatasetResponse(
