+++ src/fides/api/task/graph_task.py
@@ -7,6 +7,7 @@ from functools import wraps
 from time import sleep
 from typing import Any, Callable, Dict, List, Optional, Tuple, Union
 
+from fideslang.validation import FidesKey
 from loguru import logger
 from sqlalchemy.orm import Session
 
@@ -21,6 +22,7 @@ from fides.api.common_exceptions import (
 )
 from fides.api.graph.config import (
     ROOT_COLLECTION_ADDRESS,
+    TERMINATOR_ADDRESS,
     CollectionAddress,
     Field,
     FieldAddress,
@@ -42,6 +44,7 @@ from fides.api.models.privacy_request import (
     ExecutionLogStatus,
     PrivacyRequest,
     RequestTask,
+    TraversalDetails,
 )
 from fides.api.schemas.policy import ActionType
 from fides.api.service.connectors.base_connector import BaseConnector
@@ -64,6 +67,12 @@ from fides.config import CONFIG
 COLLECTION_FIELD_PATH_MAP = Dict[CollectionAddress, List[Tuple[FieldPath, FieldPath]]]
 
 EMPTY_REQUEST = PrivacyRequest()
+EMPTY_REQUEST_TASK = RequestTask()
+
+ARTIFICIAL_NODES: List[CollectionAddress] = [
+    ROOT_COLLECTION_ADDRESS,
+    TERMINATOR_ADDRESS,
+]
 
 
 def retry(
@@ -233,9 +242,7 @@ class GraphTask(ABC):  # pylint: disable=too-many-instance-attributes
         return f"{type(self)}:{self.key}"
 
     def generate_dry_run_query(self) -> Optional[str]:
-        """Type-specific query generated for this traversal_node.
-        TODO unaddressed -
-        """
+        """Type-specific query generated for this traversal_node."""
         return self.connector.dry_run_query(self.execution_node)
 
     def can_write_data(self) -> bool:
@@ -641,16 +648,24 @@ class GraphTask(ABC):  # pylint: disable=too-many-instance-attributes
 def collect_queries(
     traversal: Traversal, resources: TaskResources
 ) -> Dict[CollectionAddress, str]:
-    """Collect all queries for dry-run
-
-    TODO This will fail now
-    """
+    """Collect all queries for dry-run"""
 
     def collect_queries_fn(
         tn: TraversalNode, data: Dict[CollectionAddress, str]
     ) -> None:
         if not tn.is_root_node():
-            data[tn.address] = GraphTask(tn, resources).generate_dry_run_query()  # type: ignore
+            collection_data = json.loads(tn.node.collection.json())
+            # Mock a RequestTask object in memory
+            resources.privacy_request_task = RequestTask(
+                collection_address=tn.node.address.value,
+                dataset_name=tn.node.address.dataset,
+                collection_name=tn.node.address.collection,
+                collection=collection_data,
+                traversal_details=_format_traversal_details_for_save(
+                    tn.node.address, {tn.node.address: tn}
+                ),
+            )
+            data[tn.address] = GraphTask(resources).generate_dry_run_query()  # type: ignore
 
     env: Dict[CollectionAddress, str] = {}
     traversal.traverse(env, collect_queries_fn)
@@ -782,3 +797,28 @@ def build_consent_dataset_graph(datasets: List[DatasetConfig]) -> DatasetGraph:
             )
 
     return DatasetGraph(*consent_datasets)
+
+
+def _format_traversal_details_for_save(
+    node: CollectionAddress, env: Dict[CollectionAddress, TraversalNode]
+) -> Dict:
+    """Format selected TraversalNode details in a way they can be saved in the database.
+
+    This will let us execute the node when ready without having to reconstruct the traversal node later.
+    """
+    if node in ARTIFICIAL_NODES:
+        return {}
+
+    traversal_node: TraversalNode = env[node]
+    connection_key: FidesKey = traversal_node.node.dataset.connection_key
+
+    return TraversalDetails(
+        dataset_connection_key=connection_key,
+        incoming_edges=[
+            [edge.f1.value, edge.f2.value] for edge in traversal_node.incoming_edges()
+        ],
+        outgoing_edges=[
+            [edge.f1.value, edge.f2.value] for edge in traversal_node.outgoing_edges()
+        ],
+        input_keys=[tn.value for tn in traversal_node.input_keys()],
+    ).dict()
